Video tweaks and refactoring
