+++ src/render/Ease.tsx
@@ -24,6 +24,8 @@ interface IEaseDispatchProps {
 export type IEaseProps = IEaseStoreProps & IEaseDispatchProps;
 
 export class Ease extends React.Component<IEaseProps, {}> {
+    private static readonly POSTER = __dirname + "/data/heart.gif";
+
     private videoPath: string;
     private hostID: string;
     private renderedPage: JSX.Element;
@@ -32,7 +34,6 @@ export class Ease extends React.Component<IEaseProps, {}> {
         super(props);
         this.videoPath = null;
         this.hostID = null;
-
         this.props.setID(Guid.raw());
     }
 
@@ -81,6 +82,7 @@ export class Ease extends React.Component<IEaseProps, {}> {
                 this.renderedPage = (
                     <VideoHostPageContainer
                         videoSource={this.videoPath}
+                        poster=""
                     />
                 );
                 break;
@@ -88,6 +90,7 @@ export class Ease extends React.Component<IEaseProps, {}> {
             case Page.VIDEO_CLIENT:
                 this.renderedPage = (
                     <VideoClientPageContainer
+                        poster={Ease.POSTER}
                         videoSource=""
                     />
                 );
